backend/requirements.txt
# FastAPI and server
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6  # PostgreSQL driver

# Auth and security
python-jose>=3.3.0  # For JWT tokens
passlib>=1.7.4      # Password hashing
bcrypt>=4.0.1       # Password hashing
pydantic>=1.10.7
pydantic[email]>=1.10.7  # For email validation
pydantic-settings  # For settings management

# Date handling
python-dateutil>=2.8.2

# CORS
fastapi-cors>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.3.1
httpx>=0.24.0       # For testing HTTP endpoints

# Development tools
black>=23.3.0       # Code formatting
isort>=5.12.0       # Import sorting
flake8>=6.0.0       # Linting